refactor: コンポーネントごとのディレクトリ構成に変更
src/components/ をフラット配置 (Foo.tsx + Foo.css が同階層) から
コンポーネント単位のディレクトリ配置 (Foo/{Foo.tsx, Foo.css, index.ts}) に変更。
各ディレクトリには barrel index.ts を配置し、外部から `from '.../components/Foo'`
で従来通り import できるよう維持。

対象 16 コンポーネント:
AuroraLoader, BlurText, Contact, Footer, GlitchText, GooeyNav, Header,
HomePage, MySkills, Particles, Profile, SectionTitle, Service, TopFv,
ViewportHandler, WorkItem

ファイル移動には git mv を使用し履歴を保持。

import path 調整
- Header → ../GooeyNav
- TopFv → ../GlitchText, ../AuroraLoader, ../ViewportHandler, ../Particles, ../../hooks/useViewport
- HomePage → ../SectionTitle, ../WorkItem, ../Profile, ../MySkills, ../Service, ../../data/works
- SectionTitle → ../BlurText
- Contact → ../SectionTitle

備考
- TopFv 配下の Aurora.css は legacy 命名 (本来 TopFv.css に揃えるべき) だが
  本 PR では構成変更のみに留めるため命名は維持

検証
- yarn typecheck 通過
- yarn build 成功 (754ms, bundle 131.03 KB gzip — 構成変更前と完全同一)
- yarn dev 起動・配信確認
