+++ src/components/GlitchText/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./GlitchText";
