+++ src/components/Header/Header.tsx
@@ -1,6 +1,6 @@
 import React, { useState } from "react";
 import styled from "styled-components";
-import GooeyNav from './GooeyNav';
+import GooeyNav from '../GooeyNav';
 import './Header.css';
 
 const HeaderContainer = styled.header`
